+++ src/calculateTotal/generateRollTotals.ts
@@ -1,0 +1,5 @@
+import { randomNumber } from 'utils'
+
+export function generateRollTotals(sides: number, rolls: number){
+  return Array.from(Array(rolls)).map(() => randomNumber(sides))
+}
